bing operands back to 32 bit

+++ main.c
@@ -12,24 +12,24 @@ const int64_t psi[NUM_Q] = {485176247, 429380462, 1026271992, 250036504, 3787169
 const int64_t psi_neg[NUM_Q] = {320928324, 491001446, 874272385, 284619963, 371608897, 730652244};
 const int64_t n_neg[NUM_Q] = {1068303601, 1068958801, 1070465761, 1071252001, 1072234801, 1073217601};
 
-void nwc_naive(int64_t, int64_t*, int64_t*, int64_t*);
-void nwc_ntt(int64_t, int64_t*, int64_t*, int64_t, int64_t*, int64_t*, int64_t*);
-void ntt(int64_t, int64_t*, int64_t*);
-void intt(int64_t, int64_t, int64_t*, int64_t*);
+void nwc_naive(int64_t, int32_t*, int32_t*, int32_t*);
+void nwc_ntt(int64_t, int64_t*, int64_t*, int64_t, int32_t*, int32_t*, int32_t*);
+void ntt(int64_t, int64_t*, int32_t*);
+void intt(int64_t, int64_t, int64_t*, int32_t*);
 int64_t* brv_powers(int64_t, int64_t, int64_t);
 int64_t brv(int64_t, int64_t);
 
-int64_t mod(int64_t a, int64_t q) {
+int32_t mod(int64_t a, int64_t q) {
     int64_t r = a % q;
-    return r < 0 ? r + q : r;
+    return (int32_t) r < 0 ? r + q : r;
 }
 
 int main(int argc, char const *argv[]) {
     int k = 0;
-    int64_t* a = malloc(N * sizeof(int64_t));
-    int64_t* b = malloc(N * sizeof(int64_t));
-    int64_t* c1 = malloc(N * sizeof(int64_t));
-    int64_t* c2 = malloc(N * sizeof(int64_t));
+    int32_t* a = malloc(N * sizeof(int32_t));
+    int32_t* b = malloc(N * sizeof(int32_t));
+    int32_t* c1 = malloc(N * sizeof(int32_t));
+    int32_t* c2 = malloc(N * sizeof(int32_t));
     int64_t q = modulus[k];
     int64_t psi_p = psi[k];
     int64_t psi_n = psi_neg[k];
@@ -57,10 +57,10 @@ int main(int argc, char const *argv[]) {
 /**
  * Calculates the NWC of a and b with modulus q naively in O(n^2) and saves it in c.
  */
-void nwc_naive(int64_t q, int64_t* a, int64_t* b, int64_t* c) {
+void nwc_naive(int64_t q, int32_t* a, int32_t* b, int32_t* c) {
     for (int i = 0; i < N; i++) {
         for (int j = 0; j < N; j++) {
-            int64_t c_k = mod(a[i] * b[j], q);
+            int64_t c_k = mod((u_int64_t) a[i] * b[j], q);
             if (i + j < N) c[i + j] = mod(c_k + c[i + j], q);
             else c[i + j - N] = mod(c[i + j - N] - c_k, q);
         }
@@ -71,10 +71,10 @@ void nwc_naive(int64_t q, int64_t* a, int64_t* b, int64_t* c) {
  * Calculates the NWV of a and b with modulus q, tables psi and psi^-1 and multiplicative inverse of n
  * and saves it in c.
  */
-void nwc_ntt(int64_t q, int64_t* psis, int64_t* psi_ns, int64_t n_neg, int64_t* a, int64_t* b, int64_t* c) {
+void nwc_ntt(int64_t q, int64_t* psis, int64_t* psi_ns, int64_t n_neg, int32_t* a, int32_t* b, int32_t* c) {
     ntt(q, psis, a);
     ntt(q, psis, b);
-    for (int i = 0; i < N; i++) c[i] = mod(a[i] * b[i], q);
+    for (int i = 0; i < N; i++) c[i] = mod((int64_t) a[i] * b[i], q);
     intt(n_neg, q, psi_ns, c);
 }
 
@@ -82,7 +82,7 @@ void nwc_ntt(int64_t q, int64_t* psis, int64_t* psi_ns, int64_t n_neg, int64_t* 
  * Calculates the forward NTT of a with modulus q and bit-reverse ordered list psis of powers of
  * a 2n-th root of unity. The result is saved in a int bit-reversed order.
  */
-void ntt(int64_t q, int64_t* psis, int64_t* a) {
+void ntt(int64_t q, int64_t* psis, int32_t* a) {
     int64_t t = N;
     for (int m = 1; m < N; m *= 2) {
         t /= 2;
@@ -92,7 +92,7 @@ void ntt(int64_t q, int64_t* psis, int64_t* a) {
             int64_t w = psis[m + i];
             for (int j = j_1; j <= j_2; j++) {
                 int64_t U = a[j];
-                int64_t V = w * a[j + t];
+                int64_t V = (int64_t) w * a[j + t];
                 a[j] = mod(U + V, q);
                 a[j + t] = mod(U - V, q);
             }
@@ -105,7 +105,7 @@ void ntt(int64_t q, int64_t* psis, int64_t* a) {
  * ordered list psis of powers of the multiplicative inverse of a 2n-th root of unity. The result is saved in a 
  * in normal order.
  */
-void intt(int64_t inv_n, int64_t q, int64_t* psis, int64_t* a) {
+void intt(int64_t inv_n, int64_t q, int64_t* psis, int32_t* a) {
     int64_t t = 1;
     for (int m = N; m > 1; m /= 2) {
         int64_t j_1 = 0;
